Make authViewModel a dependency and separate AppIntent logic
