+++ Basic-Car-Maintenance/Shared/Localizable.xcstrings
@@ -1569,7 +1569,7 @@
       }
     },
     "Date" : {
-      "comment" : "Date picker label",
+      "comment" : "Date picker label\nThe date when the reading should be logged.",
       "localizations" : {
         "be" : {
           "stringUnit" : {
@@ -1817,6 +1817,7 @@
       }
     },
     "Distance" : {
+      "comment" : "The distance value",
       "localizations" : {
         "de" : {
           "stringUnit" : {
@@ -2737,6 +2738,9 @@
     "Imperial" : {
       "comment" : "Imperial unit system"
     },
+    "In which distance unit would you like to save the entered value?" : {
+
+    },
     "It's open source and anyone can contribute to it." : {
       "comment" : "Tells the user they can contribute to the codebase.",
       "localizations" : {
@@ -3522,9 +3526,6 @@
         }
       }
     },
-    "No vehicles available, please add a vehicle using the app and try again" : {
-      "comment" : "an error shown when attempting to add an odometer while there are no vehicles added"
-    },
     "Notes" : {
       "comment" : "Maintenance event notes text field label",
       "localizations" : {
@@ -3877,6 +3878,9 @@
     },
     "Plate: %@" : {
 
+    },
+    "Please add a distance of at least 1 kilometer or mile." : {
+      "comment" : "an error shown when entering a zero or negative value for distance"
     },
     "Preferred System" : {
 
@@ -4609,6 +4613,9 @@
           }
         }
       }
+    },
+    "Sorry, there're no vehicles saved to add a reading, please make sure you've saved at least one vehicle. You can do this in the app and then try adding the reading again." : {
+      "comment" : "an error shown when attempting to add an odometer while there are no vehicles added"
     },
     "Tap the + to begin" : {
       "localizations" : {
@@ -5381,7 +5388,7 @@
       }
     },
     "Vehicle" : {
-      "comment" : "Maintenance event vehicle picker header",
+      "comment" : "Maintenance event vehicle picker header\nThe selected vehicle to add the odometer reading to.",
       "localizations" : {
         "be" : {
           "stringUnit" : {
@@ -6322,6 +6329,9 @@
         }
       }
     },
+    "Which vehicle would you like to add this to?" : {
+
+    },
     "Year" : {
       "localizations" : {
         "be" : {
@@ -6401,9 +6411,6 @@
           }
         }
       }
-    },
-    "You can not select distance number less than 1 km or mi" : {
-      "comment" : "an error shown when entering a zero or negative value for distance"
     },
     "your vehicle" : {
       "localizations" : {
